Refactor Shutdown.py -> shutdown.py
