+++ medusa/server/web/home/handler.py
@@ -37,7 +37,7 @@ from ....search.queue import (
 )
 from ....show.Show import Show
 from ....system.Restart import Restart
-from ....system.Shutdown import Shutdown
+from ....system.shutdown import Shutdown
 from ....version_checker import CheckVersion
